+++ src/migrations/m240915_120000_convert_draft_status.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace putyourlightson\campaign\migrations;
+
+use craft\db\Migration;
+use putyourlightson\campaign\records\SendoutRecord;
+
+class m240915_120000_convert_draft_status extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function safeUp(): bool
+    {
+        $this->update(SendoutRecord::tableName(), ['sendStatus' => 'unsent'], ['sendStatus' => 'draft']);
+
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown(): bool
+    {
+        echo self::class . " cannot be reverted.\n";
+
+        return false;
+    }
+}