+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Improved the status colours of element types.
 - Improved the French translation ([#484](https://github.com/putyourlightson/craft-campaign/issues/484)).
 - Improved the German translation.
+- Renamed the “draft” sendout status to “unsent”.
 
 ### Fixed